# RedisConnection::WriteRESP: Use one asio::async_write instead of many