This increases the performance of our Redis connection and by that our Icinga DB dump speed.


### Results (Icinga DB dump times)
- Master (160d0ea3): 18s, 19s, 19s
- This PR: 14s, 14s, 14s

 
### Config

50k Hosts
50k Services
250k Customvars

```
global.numHosts = 50000
global.activeChecks = false

global.dummyResultFunction = function(checkable) {
        if (checkable.last_check_result && checkable.last_check_result.exit_status) {
                return 0
        } else {
                return 2
        }
}

global.customVarFunction = function(id) {
        var customVars = {}

        customVars.derp = id
        customVars.merp = id
        customVars.kerp = id
        customVars.mmmm = id
        customVars.perp = {}

        var numPerpVars = 50
        for (i in range(numPerpVars)) {
                customVars.perp["var" + i] = i + "_" + id
        }

        return customVars
}

for (hid in range(global.numHosts)) {
        object Host "TestHost1 - " + hid use (hid) {
                check_command = "dummy"
                vars.dummy_state = {{ global.dummyResultFunction(this) }}
                max_check_attempts = 1
                enable_active_checks = global.activeChecks
                check_interval = 5m

                vars += global.customVarFunction(hid)
        }
        object Service "TestService - " + hid use (hid) {
                host_name = "TestHost1 - " + hid
                check_command = "dummy"
                vars.dummy_state = {{ global.dummyResultFunction(this) }}
                max_check_attempts = 1
                enable_active_checks = global.activeChecks
                check_interval = 5s

                vars += global.customVarFunction(hid)
        }
}
```

closes #8692